format_path: cache project roots per working directory
The project root is discovered and cached in .onLoad().  This approach
of using one global root means that the project value doesn't follow
changes in the working directory of the session, which makes testing
the "proj" case difficult and leads to confusing behavior if the user
manually changes the working directory of the session.

Instead introduce a helper that finds the project root for the current
working directory and caches that for the session.  Also switch
format_path() over to the already document behavior of aborting when
"proj" is specified but a project isn't found, adjusting most of the
tests to use a local mrggsave.path.type="none" value rather than the
"proj" default.