mrggsave_common: continue to accept list for envir under glue 1.8.0
mrggsave() has an envir argument that is eventually passed as glue's
.envir argument.  glue 1.8.0 started aborting if the .envir argument
is not an environment.

To avoid breaking existing mrggsave callers, continue to support a
list by passing envir to glue_data's .x argument.  (An equivalent
option, provided the same parent is specified to list2env, would be to
stick with the glue::glue call and convert a list argument into an
environment.)

There is one scenario where this change leads to a change in behavior:
a caller passes a list _and_ expects to be able to look up values in
the global environment.  That old behavior arose from how glue
versions before 1.8.0 processed a list value for .envir, which led to
eval using the glue:::identity_transformer environment as the
enclosing environment (and from which the global environment is
accessible).

It seems unlikely that any callers were relying on the above behavior
(and, in any case, it's not accessible with glue 1.8.0).  For the
updated calls, specify an empty environment to get a stricter
evaluation environment when a list is passed.  If callers want to look
up values the global environment, they can leave envir at its default
value or pass an appropriate environment.